requirements.txt
streamlit
panda
numpy 
matplotlib
prophet
joblib
requests
